+++ crates/sase_core_py/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- *(model_route)* bind size-alias and epic-land model routing primitives
+
 ## [0.27.4](https://github.com/sase-org/sase-core/compare/v0.27.3...v0.27.4) - 2026-08-15
 
 ### Added